requirements.txt
git+https://github.com/huggingface/accelerate@b0e5fd353c884d443aacffa18328971db9dd728a
bitsandbytes==0.43.1
datasets==3.2.0
huggingface-hub==0.28.1
ipykernel==6.29.5
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
peft==0.11.1
python-dotenv==1.0.1
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
tokenizers==0.20.1
torch==2.5.1
transformers==4.46.1
tqdm==4.66.4
trl==0.15.2
vllm==0.6.4.post1
wandb==0.18.3